Changed input for button read to 1 Khz
Due to bounce on the button and the count increasing by many more than 1 when the button is pushed as seen during testing, a change of 1Khz might fix this issue from the current 100Mhz. Should still be fast enough to read the button push but hopefully will only increment count by 1 or at least less than seen now.